feat: add windows support (#119)
- refactor: the codebase to use `substr` instead of `range`.
- refactor: CMake targets.
- refactor: add bazelrc to build using either abseil or std::string STD.
- ci: build and run on windows using CMake and bazel.
- update: add platform support section and update building instructions.

+++ src/datadog/sampling_util.h
@@ -35,7 +35,8 @@ inline std::uint64_t max_id_from_rate(Rate rate) {
     return std::numeric_limits<uint64_t>::max();
   }
 
-  return rate * static_cast<double>(std::numeric_limits<std::uint64_t>::max());
+  return static_cast<std::uint64_t>(
+      rate * static_cast<double>(std::numeric_limits<std::uint64_t>::max()));
 }
 
 }  // namespace tracing
